6-source/requirements.txt
fake
kafka-python
pandas
openpyxl
mmh3